+++ quill/src/handlers/RotatingFileHandler.cpp
@@ -38,12 +38,6 @@ void RotatingFileHandler::write(fmt::memory_buffer const& formatted_log_record,
 /***/
 void RotatingFileHandler::_rotate()
 {
-  if (_current_index >= _backup_count)
-  {
-    // we can not rotate anymore, do nothing
-    return;
-  }
-
   if (_file)
   {
     // close the previous file
@@ -64,18 +58,23 @@ void RotatingFileHandler::_rotate()
   {
     filename_t const previous_file = detail::file_utilities::append_index_to_filename(_filename, i);
     filename_t const new_file = detail::file_utilities::append_index_to_filename(_filename, i + 1);
-
     quill::detail::rename(previous_file, new_file);
   }
 
-  // then we will always rename the base filename to 1
-  filename_t const previous_file = _filename;
-  filename_t const new_file = detail::file_utilities::append_index_to_filename(_filename, 1);
+  if (_backup_count > 0)
+  {
+    // then we will always rename the base filename to 1
+    filename_t const previous_file = _filename;
+    filename_t const new_file = detail::file_utilities::append_index_to_filename(_filename, 1);
+    quill::detail::rename(previous_file, new_file);
 
-  quill::detail::rename(previous_file, new_file);
-
-  // Increment the rotation index
-  ++_current_index;
+    // Increment the rotation index, -1 as we are counting _current_index from 0
+    if (_current_index < (_backup_count - 1))
+    {
+      // don't increment past _backup_count
+      ++_current_index;
+    }
+  }
 
   // Now reopen the base filename for writing again
   _file = detail::file_utilities::open(_filename, "w");