+++ quill/test/QuillLogTest.cpp
@@ -459,7 +459,7 @@ TEST_CASE("log_using_multiple_stdout_formats")
     if (i % 2 == 0)
     {
       std::string expected_string =
-        "QuillLogTest.cpp:340         LOG_INFO      root         - Hello log num " + std::to_string(i);
+        "QuillLogTest.cpp:436         LOG_INFO      root         - Hello log num " + std::to_string(i);
 
       if (!quill::testing::file_contains(result_arr, expected_string))
       {
@@ -470,7 +470,7 @@ TEST_CASE("log_using_multiple_stdout_formats")
     {
 
       std::string expected_string =
-        "custom - Hello log num " + std::to_string(i) + " (_DOCTEST_ANON_FUNC_14)";
+        "custom - Hello log num " + std::to_string(i) + " (_DOCTEST_ANON_FUNC_16)";
 
       if (!quill::testing::file_contains(result_arr, expected_string))
       {
@@ -500,8 +500,8 @@ TEST_CASE("log_using_stderr")
   std::string results = quill::testing::GetCapturedStderr();
 
   REQUIRE_EQ(results,
-             "log_using_stderr - Hello log stderr (_DOCTEST_ANON_FUNC_18)\n"
-             "log_using_stderr - Hello log stderr again (_DOCTEST_ANON_FUNC_18)\n");
+             "log_using_stderr - Hello log stderr (_DOCTEST_ANON_FUNC_20)\n"
+             "log_using_stderr - Hello log stderr again (_DOCTEST_ANON_FUNC_20)\n");
 }
 
 /***/