+++ CHANGELOG.md
@@ -46,6 +46,8 @@
 - Previously when the max limit of rotated files in `RotatingFileHandler` was reached, quill would stop rotating and
   instead keep logging everything into the last log file. Now when the maximum limit of files is reached,
   quill will now keep rotating by replacing the oldest logs. ([#157](https://github.com/odygrd/quill/issues/157))
+- Improve the backend logging thread responsiveness when variables are logged in loops without any delay
+  from multiple threads. ([#116](https://github.com/odygrd/quill/issues/116))
 
 ## v1.6.3